pre-release cleanup of ubuntu 16.04 build/test/.deb
