+++ .travis.yml
@@ -15,7 +15,7 @@ compiler:
 #      - cmake-data
 #      - cmake
 before_install:
-  - sudo apt-get update && sudo apt-get install -y build-essential cmake git zlib1g zlib1g-dev libssl-dev libcurl3 libcurl4-openssl-dev
+  - sudo apt-get update && sudo apt-get install -y build-essential cmake git zlib1g zlib1g-dev libssl-dev libcurl3 libcurl3-openssl-dev
   - echo $LANG
   - echo $LC_ALL
 script:
