+++ headers.h
@@ -2,7 +2,3 @@
 #include <pwd.h>
 #include <grp.h>
 #include <shadow.h>
-
-//extern "C" unsigned long hash(const char *str);
-//extern "C" enum nss_status get_posix_iam_user(char *buffer, int buflen, struct passwd *p);
-
