+++ integration-tests/sshd/Dockerfile.ubuntu-16-04
@@ -1,0 +1,26 @@
+FROM ubuntu:16.04
+
+ENV DEBIAN_FRONTEND=noninteractive
+ENV TZ=UTC
+
+RUN apt-get update      \
+ && apt-get install -y  \
+      openssh-server    \
+      sudo              \
+      awscli            \
+      curl              \
+      git               \
+      jq                \
+      vim
+
+RUN apt-get update      \
+ && apt-get install -y  \
+ zlib1g                 \
+ libssl1.0              \
+ libcurl3
+
+RUN mkdir /var/run/sshd
+
+EXPOSE 22
+
+CMD ["/opt/sshd-entrypoint.sh"]
