+++ pam.h
@@ -1,0 +1,9 @@
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+int iam_authenticate(char *user, char *pass, char *token);
+
+#ifdef __cplusplus
+}
+#endif
